Update README.md
This project was created for ES-4, Digital Logic Circuits, at Tufts University.

This project is a snake game built through sequential logic in VHDL. Players will interact with a snake (green block) and an apple (red block). Snake moves within 640 x 480 pixel display and it increases its length as it eats apples. We used a NES controller for input, and displayed output to 640x480 VGA. Code was written in Radiant and programmed to iCE40UP5k-SG48I FPGA.